database_api_docker/requirements.txt
alpaca-py==0.31.0
questdb[dataframe]~=2.0
pytest~=8.0
schedule~=1.2
fastapi[all]~=0.115
pandas~=2.2
